Clean sockets/index.js Rename sockets folder to events Create utils folder and add all utility functions
