+++ src/utils/get_winner.js
@@ -1,0 +1,9 @@
+const db = require('../storage');
+
+async function getWinner(client){
+    return await db.gameStorage.getWinner([...client.rooms][0])
+}
+
+module.exports = {
+    getWinner
+}
